+++ src/Musurp/Neo/Cypher/Builder/Statement/StatementSegment.php
@@ -13,23 +13,23 @@ namespace Musurp\Neo\Cypher\Builder\Statement;
 
 use Musurp\Neo\Cypher\Builder\BuilderInterface;
 use Musurp\Neo\Cypher\Builder\QueryBuilder;
-use Musurp\Neo\Cypher\Component\Clause\WithClause;
+use Musurp\Neo\Cypher\Component\AbstractExitClause;
 
 class StatementSegment implements BuilderInterface
 {
     private $builder;
-    private $with;
+    private $exit;
 
     /**
      * Constructor.
      *
      * @param QueryBuilder $builder
-     * @param WithClause $with
+     * @param AbstractExitClause $exit
      */
-    public function __construct(QueryBuilder $builder, WithClause $with)
+    public function __construct(QueryBuilder $builder, AbstractExitClause $exit)
     {
         $this->builder = $builder;
-        $this->with = $with;
+        $this->exit = $exit;
     }
 
     /**
@@ -39,7 +39,7 @@ class StatementSegment implements BuilderInterface
     {
         return implode(PHP_EOL, [
             $this->builder->build(),
-            $this->with->toString(),
+            $this->exit->toString(),
         ]);
     }
 }