+++ src/Musurp/Neo/Cypher/Builder/StatementBuilder.php
@@ -12,6 +12,7 @@ declare(strict_types=1);
 namespace Musurp\Neo\Cypher\Builder;
 
 use Musurp\Neo\Cypher\Builder\Statement\StatementSegment;
+use Musurp\Neo\Cypher\Component\Clause\ReturnClause;
 use Musurp\Neo\Cypher\Component\Clause\WithClause;
 
 class StatementBuilder implements BuilderInterface
@@ -22,6 +23,13 @@ class StatementBuilder implements BuilderInterface
     public function with(QueryBuilder $builder, array $variables): self
     {
         $this->segments[] = new StatementSegment($builder, new WithClause($variables));
+
+        return $this;
+    }
+
+    public function return(QueryBuilder $builder, array $variables): self
+    {
+        $this->segments[] = new StatementSegment($builder, new ReturnClause($variables));
 
         return $this;
     }