+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-- Nothing.
+- The `StatementBuilder` now supports `RETURN`.
+- The `RETURN` and `WITH` clauses now extend a common `AbstractExitClause`.
 
 ## [0.1.1] - 2018-01-25